Refactored dataLoaders, implemented inheritance for Like associated entities
