+++ src/entities/User.ts
@@ -14,6 +14,7 @@ import { Usercategory } from "./Usercategory";
 import { Charity } from "./Charity";
 import { Charityrolelink } from "./Charityrolelink"
 import { Eventvolunteer } from "./Eventvolunteer";
+import { Eventlike } from "./Eventlike";
 
 @ObjectType()
 @Entity()
@@ -38,6 +39,9 @@ export class User extends BaseEntity {
 
   @OneToMany(() => Like, (like) => like.user)
   likes: Like[];
+
+  @OneToMany(() => Eventlike, (ev) => ev.user)
+  eventlikes: Eventlike[];
 
   @OneToMany(() => Usercategory, (uc) => uc.user)
   usercategories: Usercategory[];
