+++ src/types.ts
@@ -1,9 +1,9 @@
 import {Request, Response} from 'express'
 import { Session, SessionData } from "express-session";
 import { Redis } from 'ioredis'
-import { createCategoryLoader } from './utils/createInterestsLoader';
-import { createLikesLoader } from './utils/createLikesLoader';
-import { createUserLoader } from './utils/createUserLoader';
+import { createCategoryLoader } from './utils/dataloaders/createInterestsLoader';
+import { createLikesLoader } from './utils/dataloaders/createLikesLoader';
+import { createUserLoader } from './utils/dataloaders/createUserLoader';
 
 export type MyContext = {
   req: Request & {session: Session & Partial<SessionData> & {userId?: number}};
